OCR initial changes
