+++ src/main/python/main.py
@@ -11,6 +11,8 @@ from controller.icon_controller import Icon_Controller
 from controller.user_controller import User_Controller
 from controller.friends_controller import Friends_Controller
 from utility.recurring_frequency_utility import Recurring_Frequency_Utility
+from controller.ocr_controller import OCR_Controller
+
 import schedule
 import threading
 import time
@@ -38,6 +40,7 @@ Recurring_Frequency_Controller(app)
 Icon_Controller(app)
 User_Controller(app)
 Friends_Controller(app)
+OCR_Controller(app)
 
 # # Define scheduler functions
 # def scheduler_task(flask_app):
